pyproject.toml
[project]
name = "asr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cuda-python>=12.3",
    "nemo-toolkit[asr]>=2.4.0",
    "onnx==1.18.0",
    "webvtt-py>=0.5.1",
]
